requirements.txt
aniso8601==9.0.1
astroid==2.7.3
autopep8==1.5.7
bandit==1.7.0
click==8.0.1
flake8==3.9.2
Flask==2.0.1
Flask-Cors==3.0.10
Flask-JWT==0.3.2
Flask-RESTful==0.3.9
gitdb==4.0.7
GitPython==3.1.18
gunicorn==20.1.0
isort==5.9.3
itsdangerous==2.0.1
Jinja2==3.0.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
pbr==5.6.0
platformdirs==2.3.0
prometheus-client==0.11.0
prometheus-flask-exporter==0.18.2
pycodestyle==2.7.0
pyflakes==2.3.1
PyJWT==1.4.2
pylint==2.10.2
pytz==2021.1
PyYAML==5.4.1
rope==0.19.0
six==1.16.0
smmap==4.0.0
stevedore==3.4.0
toml==0.10.2
Werkzeug==2.0.1
wrapt==1.12.1
